backend/requirements.txt
Flask==2.0.1
twilio==7.0.0
python-dotenv==0.19.0
flask-sock==0.5.0
google-cloud-speech==2.0.1
flask-cors